make getRoleRef a separate package
