+++ pkg/webhooks/utils.go
@@ -94,3 +94,17 @@ func processResourceWithPatches(patch []byte, resource []byte) []byte {
 	}
 	return resource
 }
+
+func containRBACinfo(policies []*kyverno.ClusterPolicy) bool {
+	for _, policy := range policies {
+		for _, rule := range policy.Spec.Rules {
+			if len(rule.MatchResources.Roles) > 0 {
+				return true
+			}
+			if len(rule.MatchResources.ClusterRoles) > 0 {
+				return true
+			}
+		}
+	}
+	return false
+}
